job application feature v1
Users can now apply for jobs

To apply for a job, the user presses the "apply" button. This changes application state in App.js, along with currentUser.applications (stored in context). A user applies for a job within the JobCard component. After applying for a job, the component rerenders, and the button is both disabled, and reads "applied" instead.

To implement this feature, the function applyToJob in App.js sets the application state to be: { jobId: true, etc...} and currentUser.applications to be {jobId: true, etc...}. I wanted to use an object to improve run time, but I will also try refactoring this as a set