updated job application feature
Refactored code to make tracking a user's application easier to read and debug. applicationIDs is state stored in App.js, stored as a Set. Two functions, hasApplicationID and applyToJob is in App.js, and stored in CurrentUserContext. JobCard uses these two functions to allow users to apply to a job with a click of a button.